refactor(hospitality): consolidate SSE sync into single useSSESync()
Replace the four-file split (useReservationEvents, useReservationQuerySync,
useSSEStatus, useSSEEventFeed) with a single useSSESync module. Status and
event feed are now shared via React context (SSESyncProvider) instead of
module-level mutable refs. Backoff/reconnect logic is preserved unchanged.

Closes #1991
