+++ apps/hospitality/llms.txt
@@ -199,27 +199,6 @@
       export const GUEST_SEGMENTS_QUERY_KEY: unknown;
     </item>
   </file>
-  <file path="src/hooks/useReservationEvents.ts">
-    <item priority="low">
-      type ReservationEventType = "reservation:created" | "reservation:updated" | "reservation:cancelled" | "hold:created" | "hold:released" | "hold:confirmed" | "table:updated" | "guest:lapsing";
-    </item>
-    <item priority="low">
-      interface ReservationEvent {
-        type: ReservationEventType;
-        venueId: string;
-        timestamp: string;
-        data: Reservation | Table | ReservationHold | LapsingGuest[];
-      }
-    </item>
-  </file>
-  <file path="src/hooks/useReservationQuerySync.ts">
-    <item priority="medium">
-      function canShowToast(): boolean { /* ... */ }
-    </item>
-    <item priority="high">
-      export function useReservationQuerySync() { /* ... */ }
-    </item>
-  </file>
   <file path="src/hooks/useReservations.ts">
     <item priority="high">
       export const RESERVATIONS_QUERY_KEY: unknown;
@@ -228,22 +207,6 @@
       interface UseReservationsParams {
         enabled?: boolean;
       }
-    </item>
-  </file>
-  <file path="src/hooks/useSSEEventFeed.ts">
-    <item priority="high">
-      type EventFeedListener = (event: ReservationEvent) => void;
-    </item>
-    <item priority="high">
-      export function broadcastSSEEvent(event: ReservationEvent): void { /* ... */ }
-    </item>
-  </file>
-  <file path="src/hooks/useSSEStatus.ts">
-    <item priority="high">
-      type SSEStatusListener = (isConnected: boolean, error: Error | null) => void;
-    </item>
-    <item priority="high">
-      export function notifySSEStatus(isConnected: boolean, error: Error | null): void { /* ... */ }
     </item>
   </file>
   <file path="src/hooks/useTables.ts">
