+++ llms-full.txt
@@ -3038,130 +3038,10 @@
     export const GUESTS_QUERY_KEY = "guests" as const;
     export const GUEST_SEGMENTS_QUERY_KEY = "guestSegments" as const;
   </file>
-  <file path="apps/hospitality/src/hooks/useReservationEvents.ts">
-    export type ReservationEventType =
-      | "reservation:created"
-      | "reservation:updated"
-      | "reservation:cancelled"
-      | "hold:created"
-      | "hold:released"
-      | "hold:confirmed"
-      | "table:updated"
-      | "guest:lapsing";
-    export interface ReservationEvent {
-      type: ReservationEventType;
-      venueId: string;
-      timestamp: string;
-      data: Reservation | Table | ReservationHold | LapsingGuest[];
-    }
-  </file>
-  <file path="apps/hospitality/src/hooks/useReservationQuerySync.ts">
-    function canShowToast(): boolean {
-      const now = Date.now();
-      _toastTimestamps = _toastTimestamps.filter((t) => now - t < TOAST_WINDOW_MS);
-      if (_toastTimestamps.length >= TOAST_MAX) return false;
-      _toastTimestamps.push(now);
-      return true;
-    }
-    export function useReservationQuerySync() {
-      const queryClient = useQueryClient();
-      const { selectedVenueId } = useVenue();
-      const { toast } = useToast();
-    
-      const makeEvent = (
-        type: ReservationEvent["type"],
-        data: ReservationEvent["data"]
-      ): ReservationEvent => ({
-        type,
-        venueId: selectedVenueId ?? "",
-        timestamp: new Date().toISOString(),
-        data,
-      });
-    
-      const invalidateReservations = () => {
-        queryClient.invalidateQueries({ queryKey: [RESERVATIONS_QUERY_KEY] });
-      };
-    
-      const invalidateTables = () => {
-        queryClient.invalidateQueries({ queryKey: [TABLES_QUERY_KEY] });
-      };
-    
-      const result = useReservationEvents({
-        venueId: selectedVenueId ?? undefined,
-        onReservationCreated: (reservation: Reservation) => {
-          invalidateReservations();
-          broadcastSSEEvent(makeEvent("reservation:created", reservation));
-          if (canShowToast()) {
-            toast({
-              title: "New reservation",
-              description: `${reservation.guestName ?? "Guest"} — ${new Date(reservation.startTime).toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" })}`,
-              variant: "accent",
-              duration: 5000,
-            });
-          }
-        },
-        onReservationUpdated: (reservation: Reservation) => {
-          invalidateReservations();
-          broadcastSSEEvent(makeEvent("reservation:updated", reservation));
-        },
-        onReservationCancelled: (reservation: Reservation) => {
-          invalidateReservations();
-          broadcastSSEEvent(makeEvent("reservation:cancelled", reservation));
-          if (canShowToast()) {
-            toast({
-              title: "Reservation cancelled",
-              description: `${reservation.guestName ?? "Guest"}'s reservation was cancelled`,
-              variant: "error",
-              duration: 5000,
-            });
-          }
-        },
-        onHoldCreated: (hold) => {
-          broadcastSSEEvent(makeEvent("hold:created", hold));
-        },
-        onHoldReleased: (hold) => {
-          broadcastSSEEvent(makeEvent("hold:released", hold));
-        },
-        onHoldConfirmed: (reservation: Reservation) => {
-          invalidateReservations();
-          broadcastSSEEvent(makeEvent("hold:confirmed", reservation));
-        },
-        onTableUpdated: (table) => {
-          invalidateTables();
-          broadcastSSEEvent(makeEvent("table:updated", table));
-        },
-      });
-    
-      // Broadcast connection status changes to any subscribers (e.g. ActivityFeed, HomePage)
-      useEffect(() => {
-        notifySSEStatus(result.isConnected, result.error);
-      }, [result.isConnected, result.error]);
-    
-      return result;
-    }
-  </file>
   <file path="apps/hospitality/src/hooks/useReservations.ts">
     export const RESERVATIONS_QUERY_KEY = "reservations" as const;
     export interface UseReservationsParams extends ListReservationsParams {
       enabled?: boolean;
-    }
-  </file>
-  <file path="apps/hospitality/src/hooks/useSSEEventFeed.ts">
-    type EventFeedListener = (event: ReservationEvent) => void;
-    export function broadcastSSEEvent(event: ReservationEvent): void {
-      for (const listener of _feedListeners) {
-        listener(event);
-      }
-    }
-  </file>
-  <file path="apps/hospitality/src/hooks/useSSEStatus.ts">
-    type SSEStatusListener = (isConnected: boolean, error: Error | null) => void;
-    export function notifySSEStatus(isConnected: boolean, error: Error | null): void {
-      _isConnected = isConnected;
-      _error = error;
-      for (const listener of _listeners) {
-        listener(isConnected, error);
-      }
     }
   </file>
   <file path="apps/hospitality/src/hooks/useTables.ts">
@@ -14511,7 +14391,18 @@
       }),
       Text: z.object({
         variant: z.enum(["body", "caption", "label", "detail", "display"]).optional(),
-        color: z.enum(["success", "warning", "error", "accent", "primary", "secondary", "tertiary", "on-accent"]).optional(),
+        color: z
+          .enum([
+            "success",
+            "warning",
+            "error",
+            "accent",
+            "primary",
+            "secondary",
+            "tertiary",
+            "on-accent",
+          ])
+          .optional(),
         align: z.enum(["center", "left", "right"]).optional(),
         mono: z.boolean().optional(),
         truncate: z.boolean().optional(),
